requirements.txt
langchain
langgraph
langchain-openai
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
mcp
httpx